- Se agrego el archivo gitignore y se eliminaron archivos de uso local - Se movio la clase JpaResource a la carpeta resource - Se cambiaron las referencias a 'token' por 'password' en la clase del modelo de Usuario
